requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

keyboard==0.13.5
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
SpeechRecognition==3.10.1
